// packages/github-mcp-server/src/common/types.ts
export const GitHubReferenceSchema = z.object({
	ref: z.string(),
	node_id: z.string(),
	url: z.string(),
	object: z.object({
		sha: z.string(),
		type: z.string(),
		url: z.string(),
	}),
});